service/requirements.txt
annotated-types==0.7.0 ; python_version >= "3.12"
blinker==1.9.0 ; python_version >= "3.12"
click==8.1.8 ; python_version >= "3.12"
colorama==0.4.6 ; python_version >= "3.12" and platform_system == "Windows"
flask-sqlalchemy==3.1.1 ; python_version >= "3.12"
flask==3.1.0 ; python_version >= "3.12"
greenlet==3.2.1 ; python_version < "3.14" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version >= "3.12"
gunicorn==23.0.0 ; python_version >= "3.12"
itsdangerous==2.2.0 ; python_version >= "3.12"
jinja2==3.1.6 ; python_version >= "3.12"
markupsafe==3.0.2 ; python_version >= "3.12"
packaging==25.0 ; python_version >= "3.12"
psycopg2-binary==2.9.10 ; python_version >= "3.12"
pydantic-core==2.33.1 ; python_version >= "3.12"
pydantic==2.11.3 ; python_version >= "3.12"
python-decouple==3.8 ; python_version >= "3.12"
sqlalchemy==2.0.40 ; python_version >= "3.12"
typing-extensions==4.13.2 ; python_version >= "3.12"
typing-inspection==0.4.0 ; python_version >= "3.12"
werkzeug==3.1.3 ; python_version >= "3.12"
